backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
Brotli==1.1.0
certifi==2025.7.14
cffi==2.0.0
channels==4.3.1
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
daphne==4.2.1
diff-match-patch==20241021
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-humanize==0.1.2
django-import-export==4.3.14
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.59.1
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
kiwisolver==1.4.9
lxml==6.0.0
matplotlib==3.10.5
multidict==6.7.0
numpy==2.3.2
openai==1.97.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.9.0
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==6.0.0
pyphen==0.17.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tablib==3.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
twilio==9.8.7
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
yarl==1.22.0
zope.interface==8.0.1
zopfli==0.2.3.post1
